refactor(core): moved legacy/deprecated code to legacy-dir
