+++ @navikt/core/react/src/form/combobox/combobox.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryFn } from "@storybook/react-vite";
 import React, { useEffect, useMemo, useRef, useState } from "react";
-import { Alert } from "../../alert";
 import { Button } from "../../button";
 import { Chips } from "../../chips";
+import { Alert } from "../../legacy/alert";
 import { Modal } from "../../modal";
 import { VStack } from "../../primitives/stack";
 import { BodyLong } from "../../typography";
