+++ @navikt/core/react/src/form/file-upload/file-upload.stories.tsx
@@ -2,8 +2,8 @@ import type { Meta, StoryFn } from "@storybook/react-vite";
 import React, { useEffect, useState } from "react";
 import { UploadIcon } from "@navikt/aksel-icons";
 import { FileUpload, type FileUploadItemProps } from ".";
-import { Alert } from "../../alert";
 import { Button } from "../../button";
+import { Alert } from "../../legacy/alert";
 import { VStack } from "../../primitives/stack";
 import { Heading } from "../../typography";
 import { renderStoriesForChromatic } from "../../utils/renderStoriesForChromatic";
