+++ @navikt/core/react/src/legacy/alert/index.ts
