+++ @navikt/core/react/src/legacy/confirmation-panel/ConfirmationPanel.meta.ts
