+++ @navikt/core/react/src/legacy/confirmation-panel/ConfirmationPanel.test.tsx
