+++ @navikt/core/react/src/legacy/link-panel/LinkPanel.tsx
@@ -1,8 +1,8 @@
 import React, { forwardRef } from "react";
 import { ChevronRightIcon } from "@navikt/aksel-icons";
-import { Panel } from "../panel";
-import type { OverridableComponent } from "../utils-external";
-import { cl } from "../utils/helpers";
+import { Panel } from "../../panel";
+import type { OverridableComponent } from "../../utils-external";
+import { cl } from "../../utils/helpers";
 import {
   LinkPanelDescription,
   type LinkPanelDescriptionProps,
