+++ @navikt/core/react/src/legacy/link-panel/LinkPanelDescription.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from "react";
-import { BodyLong } from "../typography";
-import { cl } from "../utils/helpers";
+import { BodyLong } from "../../typography";
+import { cl } from "../../utils/helpers";
 
 export interface LinkPanelDescriptionProps extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
