+++ @navikt/core/react/src/legacy/link-panel/index.ts
