+++ @navikt/core/react/src/legacy/stepper/Step.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from "react";
-import { Label } from "../typography";
-import type { OverridableComponent } from "../utils-external";
-import { cl, composeEventHandlers } from "../utils/helpers";
+import { Label } from "../../typography";
+import type { OverridableComponent } from "../../utils-external";
+import { cl, composeEventHandlers } from "../../utils/helpers";
 import { useStepperContext } from "./context";
 
 export interface StepperStepProps extends React.AnchorHTMLAttributes<HTMLAnchorElement> {
