+++ @navikt/core/react/src/legacy/stepper/context.ts
@@ -1,4 +1,4 @@
-import { createStrictContext } from "../utils/helpers";
+import { createStrictContext } from "../../utils/helpers";
 
 interface StepperContextValue {
   onStepChange: (step: number) => void;
