+++ @navikt/core/react/src/legacy/stepper/index.ts
